feat(mcp): refactor router export and enhance PostgreSQL server error handling
