+++ libs/product-domain/backend/src/ai-manager/infrastructure/mcp/routers/index.ts
@@ -1,1 +1,1 @@
-export * from './stateless-streamable-server-router';
+export * from './stateless-streamable-router';
