+++ libs/product-domain/backend/src/ai-manager/infrastructure/mcp/routers/stateless-streamable-router.ts
